+++ dom/base/ShadowRoot.h
@@ -12,8 +12,9 @@
 #include "nsCOMPtr.h"
 #include "nsCycleCollectionParticipant.h"
 #include "nsIContentInlines.h"
+#include "nsIdentifierMapEntry.h"
 #include "nsTHashtable.h"
-#include "nsDocument.h"
+#include "nsXBLBinding.h"
 
 class nsIAtom;
 class nsIContent;
